+++ app/src/main/java/com/example/budgettracker/database/Transaction.kt
@@ -1,10 +1,12 @@
 package com.example.budgettracker.database
 
+import java.math.BigDecimal
+
 data class Transaction(
     //var expenseID: Long,
     var userID: String? = null,
     var title: String? = null,
-    var amount: Double? = null,
+    var amount: Int? = null,
     var category: String? = null,
     var date: String? = null,
     var type: String? = null
